backend/durable_func_app/requirements.txt
azure-functions==1.24.0
azure-functions-durable==1.4.0
azure-cosmos==4.14.0
azure-storage-blob==12.27.1
redis==7.0.1